convert PhotoRoll component to PackageDestination container, passive image click console log their _id
